+++ internal/pipe/docker/doc.go
@@ -1,0 +1,3 @@
+// Package docker provides a Pipe that creates and pushes Docker images and
+// manifests.
+package docker
